+++ .gitignore
@@ -1,3 +1,2 @@
 .gitignore
 olvassel.txt
-zzWebshopReadMe.txt
