+++ css/style.css
@@ -1,0 +1,254 @@
+/* barna
+#F8F3E6
+#D1917A
+#AA583A
+#5F3121
+#241611
+*/
+
+/* narancsos
+#FFF5DC
+#FED777
+#FCBA12
+#A77802
+#342809
+*/
+
+/* fakó lila-zöld
+#FFFFFF
+#D6D9B0
+#964694
+#AF64AD
+#351E34
+*/
+
+/* lila-kék-narancs
+#D4E4F7
+#236AB9
+#FBA90A
+#5A18C9
+#190934
+*/
+
+* {
+    box-sizing: border-box;
+}
+
+:root {
+    --first: #f8f3e6;
+    --second: #D1917A;
+    --third: #AA583A;
+    --fourth: #5F3121;
+    --fifth: #241611;
+    --transition: 800ms;
+    --bs-body-color-rgb: #fff;
+}
+
+html, body {
+    background-color: var(--first);
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+}
+
+.nav-color {
+    background-image: linear-gradient(to right,var(--fifth), var(--fifth), var(--fourth))!important;
+}
+
+.nav-link, .welcome-text, .navbar-brand {
+    color: var(--first)!important;
+    transition: all var(--transition);
+}
+
+.nav-link:hover {
+    color: var(--second)!important;
+    text-decoration: underline 1px solid var(--second)!important;
+}
+
+.navbar-brand:hover {
+    color: var(--first)!important;
+}
+
+
+
+.main-custom-top {
+    margin-top: 3.5rem;
+}
+
+/* tesztelés */
+@media screen and (max-width: 330px) {
+    .main-custom-top {
+        margin-top: 10rem;
+    }
+}
+
+
+
+.week-color {
+    background-color: var(--first);
+    z-index: 50!important;
+}
+
+/* dolgozni kell még rajta MINDKÉT css-be */
+.week-offer {
+    animation-name: week-offer!important;
+    animation-duration: 4s!important;
+    animation-iteration-count: infinite!important;
+}
+
+@keyframes week-offer {
+    from { opacity: 0; };
+    to { opacity: 1; };
+}
+
+.log-name {
+    color: var(--third)!important;
+}
+
+.border-card {
+    border: 1px transparent;
+}
+
+.border-custom {
+    border-bottom: 1px solid var(--fifth)!important;
+}
+
+.border-custom-thick {
+    border: 2px solid var(--fifth)!important;
+}
+
+/* .card */
+.offer-card {
+    background-color: var(--first)!important;
+    transition: all var(--transition) ease-out;
+}
+
+/* .card:hover */
+.offer-card:hover {
+    box-shadow: 0 0 1rem var(--fifth)!important;
+    transform: translateY(-.3rem);
+}
+
+.products-card {
+    background-color: var(--first)!important;
+    transform: skewY(1deg);
+    transition: all var(--transition) ease-out;
+}
+
+.products-card:hover {
+    box-shadow: 0 0 1rem var(--fifth)!important;
+    transform: skewY(0);
+}
+
+.custom-top-carousel {
+    margin-top: 2.5rem;
+}
+
+.custom-carousel-inner {
+    height: 500px;
+}
+
+.custom-carousel {
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    text-align: center;   
+    background-color: var(--third)!important;
+}
+
+.custom-carousel-height {
+    height: 100%;
+}
+
+.contact-main {
+    display: flex;
+    justify-content: center;
+}
+
+.contact-list {
+    display: flex;
+    flex-direction: column;
+    gap: .5rem;
+    width: 100%;
+    padding: 0;
+    text-align: center;
+    list-style-type: none;
+    font-weight: bold;
+}
+
+.modal-title {
+    color: var(--first)!important;
+}
+
+.termek-cim, .name-color, .text-color, body, h1, h2, h3, h4, h5, h6 {
+    color: var(--fifth)!important;
+}
+
+.modal-custom {
+    background-color: var(--fifth)!important;
+}
+
+.modal-text-custom, .date {
+    color: var(--first)!important;
+}
+
+input:focus, textarea:focus, select:focus {
+    border-color: var(--fifth)!important;
+    box-shadow: 1px 1px 1px var(--fifth) inset, 0 0 8px var(--fourth)!important;
+    outline: 0 none!important;
+}
+
+.comment_mail {
+    color: var(--fourth)!important;
+    transition: all var(--transition)!important;
+}
+
+.comment_mail:hover {
+    color: var(--fifth)!important;
+    text-decoration: 1px solid var(--fifth)!important;
+}
+
+.message_date {
+    color: var(--fourth)!important;
+}
+
+.message_text {
+    color: var(--third)!important;
+}
+
+/* fizetes.php btn-primary? active? */
+/* kosar.php törlés gombok? active? */
+
+.navbar-toggler {
+    background-color: var(--fourth)!important;
+    border-color: var(--first)!important;
+    color: var(--fifth)!important;
+}
+
+.btn-dark {
+    background-color: var(--fifth);
+    color: var(--first);
+    transition: all var(--transition);
+}
+
+.btn-dark:hover {
+    background-color: var(--fifth);
+    opacity: 0.7;
+}
+
+.btn-dark:active {
+    background-color: var(--fifth)!important;
+    transform: scale(0.9);
+}
+
+/* új bootstrap verzió után szükséges, a táblázatok fehér háttérrel rendelkeznek enélkül */
+table > thead > tr > th, table > tbody > tr > td {
+    background-color: transparent!important;
+}
+
+/* ha külön td-be lesznek akkor nem kell */
+@media screen and (min-width: 768px) {
+    .input-qtty {
+        width: 4rem;
+    }    
+}
