+++ data/payment.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Gép: localhost
--- Létrehozás ideje: 2024. Jan 06. 15:43
+-- Létrehozás ideje: 2024. Már 02. 15:40
 -- Kiszolgáló verziója: 8.0.17
 -- PHP verzió: 7.3.10
