+++ data/users.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Gép: localhost
--- Létrehozás ideje: 2024. Jan 10. 13:54
+-- Létrehozás ideje: 2024. Már 02. 15:41
 -- Kiszolgáló verziója: 8.0.17
 -- PHP verzió: 7.3.10
 
@@ -67,10 +67,18 @@ INSERT INTO `users` (`id`, `first_name`, `last_name`, `email`, `password`, `phon
 (18, 'Takács', 'Péter', 'ptakacs@gmail.com', '12344321', '+36102156987', 'Takács Péter', 'Magyarország', 9000, 'Győr', 'Nyugat utca', '19', '2023-12-12 17:51:00', NULL),
 (19, 'Vörös', 'János', 'vjanos@gmail.com', 'vjani1212', '+36124967712', 'Vörös János', 'Magyarország', 8243, 'Balatonakali', 'Balaton utca', '5', '2023-12-12 17:55:26', NULL),
 (20, 'Nagy', 'Zita', 'nzita@mail.com', 'nzita123', '+36105557846', 'Nagy Zita', 'Magyarország', 3348, 'Szilvásvárad', 'Hegyi utca', '10', '2023-12-12 17:58:19', NULL),
-(21, 'Szegedi', 'János', 'szegedij19@gmail.com', '12341121', '+36105597788', 'Szegedi János', 'Magyarország', 6500, 'Baja', 'Szegedi út', '129', '2022-05-23 15:34:48', '2023-12-11 21:17:52'),
 (22, 'Magyar', 'Péter', 'mpeter@teszt.dev', 'teszt222', '+36105551271', 'Magyar Péter', 'Magyarország', 1044, 'Budapest', 'Tésztaszűrő utca', '77', '2024-01-06 12:25:49', NULL),
 (23, 'Harang', 'Virág', 'haravirag@teszt.dev', 'harangvirag', '+36121794578', 'Harang Virág', 'Magyarország', 1022, 'Budapest', 'Ibolya utca', '17', '2024-01-06 12:52:55', NULL),
-(24, 'Ügyet', 'Lenke', 'ugyet@lenke.hu', 'ablakzsiraf', '+36105498789', 'Ügyet Lenke Irén', 'Magyarország', 1044, 'Budapest', 'Bécsi utca', '135', '2024-01-08 12:57:48', '2024-01-10 13:39:36');
+(24, 'Ügyet', 'Lenke', 'ugyet@lenke.hu', 'ablakzsiraf', '+36105498789', 'Ügyet Lenke Irén', 'Magyarország', 1044, 'Budapest', 'Bécsi utca', '135', '2024-01-08 12:57:48', '2024-01-10 13:39:36'),
+(25, 'Admin', 'Admin', 'admin@info.hu', 'admin', '+36121231234', 'Admin Admin', 'Magyarország', 1022, 'Budapest', 'Tököl utca', '19', '2024-02-15 23:56:23', NULL),
+(26, 'Magyar', 'József', 'mjozsef@info.hu', 'jelszo12', '+36124445555', 'Magyar József', 'Magyarország', 6000, 'Kecskemét', 'József Attila utca', '39', '2024-02-16 12:42:43', NULL),
+(27, 'Halasi', 'Dóra', 'dora@email.hu', '11223344', '+36124445454', 'Halasi Dóra', 'Magyarország', 1022, 'Budapest', 'Tiszta utca', '121', '2024-02-16 14:37:14', NULL),
+(28, 'Teszt', 'Elek', 'tesztelek@info.hu', '11223344', '+36124567892', 'Teszt Elek', 'Magyarország', 1023, 'Budapest', 'Szent Isvtán király utca', '21', '2024-02-16 15:16:23', '2024-02-19 13:49:31'),
+(29, 'Szabadi', 'Éva', 'szeva@gmail.com', '1234512345', '+36105497878', 'Szabadi Éva', 'Magyarország', 1024, 'Budapest', 'Titkos utca', '1', '2024-02-16 18:01:43', '2024-02-16 18:36:02'),
+(30, 'Havas', 'Mihály', 'hmisi@email.com', '111222333', '+36121256541', 'Havas Mihály', 'Magyarország', 1023, 'Budapest', 'Zöldellő utca', '127', '2024-02-16 18:44:20', NULL),
+(31, 'Tiszta', 'Tibor', 'titi@gmail.com', '123123123', '+36105469878', 'Tiszta Tibor', 'Magyarország', 1024, 'Budapest', 'Lánchíd utca', '12', '2024-02-16 18:50:15', NULL),
+(32, 'Minta', 'Polgár', 'mpolgar@mail.hu', '123321123', '+36104976523', 'Minta Polgár', 'Magyarország', 1022, 'Budapest', 'Kiss László utca', '1', '2024-02-16 18:52:59', NULL),
+(33, 'Tóth', 'Tamás', 'ttom@mail.com', '22112211', '+36120014510', 'Tóth Tamás', 'Magyarország', 1024, 'Budapest', 'Tóth Mike utca', '27', '2024-02-16 19:07:44', NULL);
 
 --
 -- Indexek a kiírt táblákhoz
@@ -90,7 +98,7 @@ ALTER TABLE `users`
 -- AUTO_INCREMENT a táblához `users`
 --
 ALTER TABLE `users`
-  MODIFY `id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=25;
+  MODIFY `id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=34;
 COMMIT;
 
 /*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
