--- logout.php
@@ -1,7 +1,0 @@
-<?php
-
-session_start();
-
-unset($_SESSION["user"]);
-
-header("location: index.php");
